--- libs/membership/membership-common/src/main/java/net/corda/membership/lib/p2p/package-info.java
@@ -1,4 +1,0 @@
-@Export
-package net.corda.membership.lib.p2p;
-
-import org.osgi.annotation.bundle.Export;
