+++ libs/membership/membership-common/src/main/kotlin/net/corda/membership/lib/grouppolicy/GroupPolicyConstants.kt
@@ -1,5 +1,9 @@
 package net.corda.membership.lib.grouppolicy
 
+import net.corda.libs.configuration.SmartConfig
+import net.corda.schema.configuration.ConfigKeys
+import net.corda.v5.base.exceptions.CordaRuntimeException
+
 /**
  * Constants for [GroupPolicy]. For example, expected keys and default values.
  */
@@ -180,6 +184,32 @@ class GroupPolicyConstants {
                 override fun toString(): String {
                     return mode
                 }
+            }
+
+            /**
+             * Enum defining the TLS type values.
+             */
+            enum class TlsType(
+                val groupPolicyName: String,
+            ) {
+                ONE_WAY("OneWay"),
+                MUTUAL("Mutual");
+
+                companion object {
+                    fun getClusterType(configurationGetter: (String)-> SmartConfig?) : TlsType {
+                        val gatewayConfiguration =
+                            configurationGetter.invoke(ConfigKeys.P2P_GATEWAY_CONFIG) ?:
+                            throw FailToReadClusterTlsTypeException(
+                                "Could not get the Gateway configuration"
+                            )
+                        val tlsType = gatewayConfiguration
+                            .getConfig("sslConfig")
+                            .getString(PolicyKeys.P2PParameters.TLS_TYPE)
+                        return valueOf(tlsType)
+                    }
+                }
+
+                class FailToReadClusterTlsTypeException(message: String)  : CordaRuntimeException(message)
             }
         }
     }
